permission issues done, url mapping done

+++ web/api/controller/comment.py
@@ -10,12 +10,12 @@ from api.model.profile import Profile
 from api.model.comment import Comment
 from api.serializer.comment import CommentSerializer
 from rest_framework.decorators import api_view, permission_classes
-from rest_framework.permissions import IsAuthenticated
-from django.views.decorators.csrf import csrf_exempt
+from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
+from api.service import permission
 
-@csrf_exempt
+
 @api_view(['GET', 'POST'])
-@permission_classes((IsAuthenticated, ))
+@permission_classes((IsAuthenticatedOrReadOnly, ))
 def comment_get_post(request):
 
     if request.method == 'GET':
@@ -25,7 +25,7 @@ def comment_get_post(request):
         except Comment.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
-    elif request.method == 'POST':
+    elif request.method == 'POST' and request.user.is_authenticated:
 
         username = request.user.username
         request.data['creator'] = Profile.objects.filter(username=username).first().pk
@@ -37,9 +37,8 @@ def comment_get_post(request):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-@csrf_exempt
-@api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes((IsAuthenticated, ))
+@api_view(['GET', 'DELETE'])
+@permission_classes((IsAuthenticatedOrReadOnly, ))
 def comment_get_put_delete(request, comment_id):
     try:
         comment = Comment.objects.get(id=comment_id)
@@ -50,21 +49,11 @@ def comment_get_put_delete(request, comment_id):
         serializer = CommentSerializer(comment)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    elif request.method == 'PUT':
-        username = request.user.username
-        request.data['creator'] = Profile.objects.filter(username=username).first().pk
-
-        serializer = CommentSerializer(instance=comment, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
-        else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    elif request.method == 'DELETE':
+    elif request.method == 'DELETE' and permission.isOwner(request, obj=comment):
         try:
             comment.delete()
             return Response(status=status.HTTP_200_OK)
         except Exception:
             return Response(status=status.HTTP_204_NO_CONTENT)
 
+    return Response(status=status.HTTP_412_PRECONDITION_FAILED)
